requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.0

# Database - PostgreSQL only (HAPUS sqlite)
psycopg2-binary==2.9.9  # PostgreSQL adapter
sqlalchemy==2.0.25
alembic==1.13.1  # Untuk migrations

# AI & Integration
ollama==0.1.8
httpx==0.26.0

# Testing
pytest==8.0.0

# PDF Processing (untuk OrderBot)
pypdf2==3.0.1

# External APIs
google-api-python-client==2.108.0
simple-salesforce==1.12.6

# Utilities
python-dotenv==1.0.0